Refactored classes into specific packages
