--- src/main/java/entities/placeholderentity.java
@@ -1,3 +1,0 @@
-package entities;
-
-// This is a placeholder entity, removed later.
