+++ src/main/java/fileio_use_case/application_business/SectionBuilder.java
@@ -1,4 +1,4 @@
-package fileio_use_case;
+package fileio_use_case.application_business;
 
 import entities.Block;
 import entities.Section;
