+++ src/test/java/feature_6/SessionGatewayInteractorTest.java
@@ -4,8 +4,8 @@ import entities.Block;
 import entities.CalendarCourse;
 import entities.Section;
 import entities.Session;
-import fileio_use_case.FileImportRequestModel;
-import fileio_use_case.SessionGatewayInteractor;
+import fileio_use_case.application_business.FileImportRequestModel;
+import fileio_use_case.application_business.session_specific_classes.SessionGatewayInteractor;
 import org.json.simple.parser.ParseException;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
